Add files via upload
Today, everything in the world is smart and digital. Many advances have also been made in the transport sector. However, public transport buses in India have always been an area where such new developments have changed their faces. Work for public transport is one of the areas of intelligent vehicle research. Every bus is controlled by a conductor. The conductor will collect money from each passenger and issue a ticket. To overcome this, we will create an IOT based ticketing system. The purpose of this project is to calculate the passenger using the IR sensor and automatically calculate the distance traveled by the passenger using the GPS sensor, and the corresponding amount is debited from the RFID card.